Implement waiting for process by name
This change implements the '-w' command line parameter to
allow waiting for a specific process name to launch.

This is done by looping over the /proc/ filesystem and
comparing the names as returned by GetProcessName to the
name passed on the command line.

GetProcessName printed error messages if errors occurred -
as we're now calling it a lot in WaitForProcessName instead
of once before, let's change the log level to 'debug' for
all messages in GetProcessName and add a new error message
which is printed when the previously existing call to
GetProcessName fails.