Add testcase for '-w'
This change adds a very simple test application to the test
suite. The application is built through the top level
Makefile and used by the *_by_name.sh scenarios.

Some changes were necessary in runProcDumpAndValidate.sh to
accommodate running ProcDump with the -w parameter:
the script checks if TESTPROGNAME is defined and not empty.
This variable is set by the new scenarios. If it is set,
the corresponding binary in the bin/ folder is run with
parameters defined in TESTPROGMODE after a small delay.