Implement waiting for process by name (-w parameter) (#34)
* Implement waiting for process by name

This change implements the '-w' command line parameter to
allow waiting for a specific process name to launch.

This is done by looping over the /proc/ filesystem and
comparing the names as returned by GetProcessName to the
name passed on the command line.

GetProcessName printed error messages if errors occurred -
as we're now calling it a lot in WaitForProcessName instead
of once before, let's change the log level to 'debug' for
all messages in GetProcessName and add a new error message
which is printed when the previously existing call to
GetProcessName fails.

* Add testcase for '-w'

This change adds a very simple test application to the test
suite. The application is built through the top level
Makefile and used by the *_by_name.sh scenarios.

Some changes were necessary in runProcDumpAndValidate.sh to
accommodate running ProcDump with the -w parameter:
the script checks if TESTPROGNAME is defined and not empty.
This variable is set by the new scenarios. If it is set,
the corresponding binary in the bin/ folder is run with
parameters defined in TESTPROGMODE after a small delay.