Consolidating all development into master branch to simplify contribution (#43)
* Add parentheses around assignments in conditionals

< and != have higher operator precedence than =

Signed-off-by: Michael Sartain <mikesart@fastmail.com>

* Fix a typo in README (with -> will)

* Fix path to sources file

* fixed. missing number of dumps

* fixed potential deadlock upon malloc failing and incorrect parsing of process name

* moving process sanitization to gcore call & adding travis CI

* fixing .travis.yml file rename

* adding dependency install steps for travis

* skip normal build steps within travis and run our own

* adding gnu flags for travis

* fixing path travis

* continued updates to travis

* gcore is part of gdb not its own package.....

* Update Procdump.c

typo fix in warning message

* adding travis CI badge

* fixed reading from /proc/PID/cmdline so that it only reads on first monitoring loop

* tests/integration: Drop dpkg/apt calls in run.sh

procdump is prefectly capable of running on more than Debian and Ubuntu,
and rather than attempting to support a bunch of different package
managers in this integration script, we'll just abort and tell the user
what they need to install for the tests to function.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* Makefile: Require gnu99 standard from the compiler

This fixes compilation on RHEL/CentOS 7, as the system GCC does not
default to C99 with GNU extensions by default. The default changed
from gnu90 to gnu11 (C11 with GNU extensiosn) in GCC 5.1.0.

Enforcing gnu99 across the board works well enough and doesn't
break building on Ubuntu, either.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* Makefile: Add DESTDIR support for 'make install'

In order to enable distributions to properly package procdump,
the Makefile needs to support DESTDIR to enable proper support
for 'make install DESTDIR=foo' used by package building tools.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* README.md: Note the zlib requirement for building the code

The 'zconf.h' header is required to build procdump, which
means that a build-time requirement for zlib headers is required.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* README.md: Update the list of minimum supported Linux distributions

This has been tested to build and run on RHEL/CentOS 7, Fedora 26+,
Mageia 6, and Ubuntu 14.04+.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* Makefile: Fix to allow packager compiler flags to be added

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* procdump.gif: Drop executable bit

Images aren't supposed to be marked as executables...

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

* moved & refactored process name extraction

* fixes to clear -Wall

* updated help statement for consistency

* made sanitize non-destructive

* fixed wrong dump desc in time thread proc

* removed unnecessary GetProcessStat call

* re-recorded hero gif

* updating version number to 1.0.1

* re-re-recorded hero gif to capture updated version number

* Update README.md

Fixed typo in usage

* Update procdump.1

Fixed typo

* Implement waiting for process by name (-w parameter) (#34)

* Implement waiting for process by name

This change implements the '-w' command line parameter to
allow waiting for a specific process name to launch.

This is done by looping over the /proc/ filesystem and
comparing the names as returned by GetProcessName to the
name passed on the command line.

GetProcessName printed error messages if errors occurred -
as we're now calling it a lot in WaitForProcessName instead
of once before, let's change the log level to 'debug' for
all messages in GetProcessName and add a new error message
which is printed when the previously existing call to
GetProcessName fails.

* Add testcase for '-w'

This change adds a very simple test application to the test
suite. The application is built through the top level
Makefile and used by the *_by_name.sh scenarios.

Some changes were necessary in runProcDumpAndValidate.sh to
accommodate running ProcDump with the -w parameter:
the script checks if TESTPROGNAME is defined and not empty.
This variable is set by the new scenarios. If it is set,
the corresponding binary in the bin/ folder is run with
parameters defined in TESTPROGMODE after a small delay.

* :memo: Adding blurb about branching to contributing doc (#42)

ship it!