ProcDumpConfiguration: fix crash in getopt_long
As reported in #92, parsing the command line parameters can
currently lead to a crash in `getopt_long`. This is caused by
a missing terminating entry in the `long_options` array
which is required as described in the man page for
getopt_long [1]. Without this entry, getopt_long continues to
parse garbage values on the stack as options.

Adding the missing entry fixes the crash.

[1]: `The last element of the array has to be filled with zeros.`

Fixes: #92