Allow custom path/file name for core dumps
This change allows to specify a custom output directory and
file name for the generated core dump files. So far, the
dumps were always places in the current directory and had the
format `<process name>_<type>_<date>.<pid>`.

Now, using the `-o` command line option, there are the
following options:

- if you pass an existing directory or a path that ends in a
  slash, we use the passed path as the target directory while
  the name stays in the default format.
- otherwise, the passed string is split up into `directory`
  and `basename` components, and the core dumps are created in
  `<directory>` with the name `<basename>_<counter>.<pid>`.

The new test cases check if using a non-existing directory
fails and if using a custom file name produces a core dump
file in the temp directory created for the test case.

Closes: #100