Fixing some leaks in src/Monitor.c code (#204)
* Monitor: Fix double-free occurence of monitoredProcessMap

GCC 13 is reporting a double free occurence of monitoredProcessMap. It
is possible by using `-Werror=analyzer-double-free`. The variable
`monitoredProcessMap` is defined as a auto_free which does not require
an explicit `free()`.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* Monitor: Fix missing var_end at GetClientDataHelper

If the `clientData` cannot be allocated, it finishes the function
earlier, in order to complete the execution properly, the variables
need to be ended with `va_end` either.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

---------

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>